+++ src/pysdccc/_cli.py
@@ -7,7 +7,6 @@ import shutil
 import subprocess
 import sys
 import tempfile
-import time
 import zipfile
 
 import httpx
@@ -17,24 +16,29 @@ from pysdccc import _common
 try:
     import click
 except ImportError as import_error:
-    msg = 'Cli not installed. Please install "pysdccc[cli]" package.'
-    raise ImportError(msg) from import_error
+    raise ImportError('Cli not installed. Please install "pysdccc[cli]" package.') from import_error  # noqa: EM101, TRY003
 
 
-class UrlType(click.ParamType):
-    """Url type."""
+class LocalOrRemotePath(click.ParamType):
+    """Local or remote path type."""
 
-    name = 'url'
+    name = 'path'
 
-    def convert(self, value: str, param: click.Parameter | None, ctx: click.Context | None) -> httpx.URL:
-        """Convert value to url."""
-        try:
-            return httpx.URL(value)
-        except Exception as e:  # noqa: BLE001
-            self.fail(f'{value} is not a valid url: {e}', param, ctx)
+    def convert(self, value: str, param: click.Parameter | None, ctx: click.Context | None) -> httpx.URL | pathlib.Path:
+        """Convert value to url or filepath."""
+        if _common.is_remote_path(value):
+            try:
+                return httpx.URL(value)
+            except Exception as e:  # noqa: BLE001
+                self.fail(f'{value} is not a valid url: {e}', param, ctx)
+        else:
+            path = pathlib.Path(value)
+            if not path.is_file():
+                self.fail(f'{value} is not a valid file path.', param, ctx)
+            return path
 
 
-URL = UrlType()
+PATH = LocalOrRemotePath()
 
 
 class ProxyType(click.ParamType):
@@ -69,21 +73,22 @@ def _download_to_stream(url: httpx.URL, stream: io.IOBase, proxy: httpx.Proxy | 
                 stream.write(chunk)
                 progress.update(response.num_bytes_downloaded - num_bytes_downloaded)
                 num_bytes_downloaded = response.num_bytes_downloaded
-                time.sleep(0.1)
 
 
-def download(url: httpx.URL, output: pathlib.Path, proxy: httpx.Proxy | None = None):
-    """Download and extract the executable from a url."""
-    click.echo(f'Downloading SDCcc from {url}.')
-    with tempfile.NamedTemporaryFile('wb', suffix='.zip', delete=False) as temporary_file:
-        _download_to_stream(url, temporary_file, proxy=proxy)  # type: ignore[arg-type]
+def extract_zip_file(zip_file: _common.PATH_TYPE, output: _common.PATH_TYPE):
     click.echo(f'Extracting SDCcc to {output}.')
     with (
-        zipfile.ZipFile(temporary_file.name) as f,
+        zipfile.ZipFile(zip_file) as f,
         click.progressbar(f.infolist(), label='Extracting', width=0) as progress,
     ):
         for member in progress:
             f.extract(member, output)
+
+
+def download(url: httpx.URL, stream: io.IOBase, proxy: httpx.Proxy | None = None):
+    """Download the executable from a url to a given stream."""
+    click.echo(f'Downloading SDCcc from {url}.')
+    _download_to_stream(url, stream, proxy=proxy)
 
 
 @click.group(help='Manage SDCcc installation.')
@@ -92,29 +97,43 @@ def cli():
     pass
 
 
-@cli.command(
-    short_help='Install the SDCcc executable from the specified URL. Releases can be found at https://github.com/Draegerwerk/SDCcc/releases.',
-)
-@click.argument('url', type=URL)
+@cli.command(short_help='Install the SDCcc executable from the specified local or remote path.')
+@click.argument('path', type=PATH)
 @click.option('--proxy', help='Proxy server to use for the download.', type=PROXY)
 @click.pass_context
-def install(ctx: click.Context, url: httpx.URL, proxy: httpx.Proxy | None):
-    """Download the specified version from the default URL to a temporary directory.
-
-    This function downloads the SDCcc executable from the given URL to a temporary directory.
-    It optionally uses a proxy and a specified timeout for the download operation.
-    The downloaded file is extracted to a local path determined by the version string in the URL.
+def install(ctx: click.Context, path: httpx.URL | pathlib.Path, proxy: httpx.Proxy | None):
+    """Install the specified version from the local or remote path.
 
     :param ctx: context from click
-    :param url: The parsed URL from which to download the executable.
+    :param path: The URL from which to download the executable or a local path from which the file is extracted.
     :param proxy: Optional proxy to be used for the download.
     """
     ctx.invoke(uninstall)
+    if isinstance(path, httpx.URL):
+        try:
+            with tempfile.NamedTemporaryFile('wb', suffix='.zip', delete=False) as stream:
+                download(path, stream, proxy)  # ty:ignore[invalid-argument-type]
+            file_to_be_extracted = stream.name
+        except Exception as e:
+            with contextlib.suppress(OSError):
+                pathlib.Path(stream.name).unlink()
+            msg = f'Failed to download SDCcc from {path}: {e}'
+            raise click.ClickException(msg) from e
+    elif isinstance(path, pathlib.Path):
+        file_to_be_extracted = str(path)
+    else:
+        msg = f'Unexpected type of path: {type(path)}'
+        raise TypeError(msg)
+
     try:
-        download(url, _common.DEFAULT_STORAGE_DIRECTORY, proxy)
+        extract_zip_file(file_to_be_extracted, _common.DEFAULT_STORAGE_DIRECTORY)
     except Exception as e:
-        msg = f'Failed to download and extract SDCcc from {url}: {e}'
+        msg = f'Failed to extract SDCcc from {path}: {e}'
         raise click.ClickException(msg) from e
+    finally:
+        if isinstance(path, httpx.URL):
+            with contextlib.suppress(OSError):
+                pathlib.Path(file_to_be_extracted).unlink()
 
 
 @cli.command(short_help='Uninstall the SDCcc executable by removing the directory.')