+++ src/pysdccc/_download.py
@@ -4,13 +4,14 @@ import concurrent.futures
 import contextlib
 import logging
 import os
-import pathlib
 import subprocess
 import sys
 import zipfile
 from collections.abc import AsyncGenerator
+from typing import cast
 
 import anyio.from_thread
+import anyio.to_thread
 import httpx
 
 from pysdccc import _common, _runner
@@ -23,14 +24,29 @@ else:
 logger = logging.getLogger('pysdccc.download')
 
 
+def _extract_zip_file_sync(zip_file_path: _common.PATH_TYPE, output: _common.PATH_TYPE) -> None:
+    """Extract the given zip file to the given output directory."""
+    with zipfile.ZipFile(zip_file_path) as f:
+        f.extractall(output)
+
+
+async def extract_zip_file(zip_file_path: _common.PATH_TYPE, output: _common.PATH_TYPE) -> None:
+    """Extract the given zip file to the given output directory.
+
+    :param zip_file_path: The path to the zip file to be extracted.
+    :param output: The path to the directory where the zip file will be extracted.
+    """
+    logger.info('Extracting SDCcc to %s.', output)
+    await anyio.to_thread.run_sync(_extract_zip_file_sync, zip_file_path, output)
+
+
 @contextlib.asynccontextmanager
 async def _open_download_stream(
     url: httpx.URL,
     proxy: httpx.Proxy | None = None,
 ) -> AsyncGenerator[httpx.Response, None]:
     """Open a stream from which SDCcc can be downloaded chunk by chunk."""
-    client = httpx.AsyncClient(follow_redirects=True, proxy=proxy)
-    async with client.stream('GET', url) as response:
+    async with httpx.AsyncClient(follow_redirects=True, proxy=proxy) as client, client.stream('GET', url) as response:
         response.raise_for_status()
         yield response
 
@@ -54,12 +70,16 @@ async def download(
         anyio.NamedTemporaryFile('wb', suffix='.zip', delete=False) as temporary_file,
         _open_download_stream(url, proxy=proxy) as response,
     ):
+        temporary_file_path = anyio.Path(cast('str', temporary_file.name))
         async for chunk in response.aiter_bytes():
             await temporary_file.write(chunk)
     output = output or _common.DEFAULT_STORAGE_DIRECTORY
-    logger.info('Extracting SDCcc to %s.', output)
-    with zipfile.ZipFile(temporary_file.name) as f:  # pyright: ignore [reportCallIssue, reportArgumentType]
-        f.extractall(output)
+    try:
+        logger.info('Extracting SDCcc to %s.', output)
+        await extract_zip_file(temporary_file_path, output)
+    finally:
+        with contextlib.suppress(OSError):
+            await temporary_file_path.unlink()
     return _common.get_exe_path(output)
 
 
@@ -71,9 +91,25 @@ async def is_downloaded(version: str) -> bool:
     :return: True if the executable is already downloaded, False otherwise.
     """
     try:
-        return await _runner.SdcccRunner(pathlib.Path().absolute()).get_version() == version
+        return await _runner.SdcccRunner(await anyio.Path().absolute()).get_version() == version
     except (FileNotFoundError, subprocess.CalledProcessError):
         return False
+
+
+async def install(path: _common.PATH_TYPE, output: _common.PATH_TYPE | None = None) -> None:
+    """Install the SDCcc executable from the given remote or local path.
+
+    This function installs the SDCcc executable from the given path. It checks if the path is a URL or a local file
+    path and calls the appropriate function to download and extract the zip file.
+
+    :param path: The remote or local path to the SDCcc executable.
+    :param output: The path to the directory where the downloaded executable will be extracted. If None,
+    `DEFAULT_STORAGE_DIRECTORY` is used.
+    """
+    if _common.is_remote_path(path):
+        await download(str(path), output=output or _common.DEFAULT_STORAGE_DIRECTORY)
+    else:
+        await extract_zip_file(path, output or _common.DEFAULT_STORAGE_DIRECTORY)
 
 
 @deprecated('Prefer using the async version `download` instead.')